// apps/web/src/lib/service-clients/service-storage/generated/schemas/propertyDefinitionWithOptions.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * document_storage_service
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyDefinition } from './propertyDefinition';
import type { PropertyOption } from './propertyOption';

/**
 * Property definition with its associated options (service representation).
 */
export interface PropertyDefinitionWithOptions {
  definition: PropertyDefinition;
  property_options: PropertyOption[];
}